allow messages of all question marks

+++ discordtest.py
@@ -473,12 +473,10 @@ async def on_message(message):
         return
     if message.content.startswith(bot.command_prefix):
         await bot.process_commands(message)
-    elif message.content.startswith('?'):
-        if len(message.content) >= 2:
-            if message.content[1] != ' ':
-                message.content = '!'+message.content[1:]
-                await bot.delete_message(message)
-                await bot.process_commands(message)
+    elif message.content.startswith('?') and not message.content.endswith('?'):
+        message.content = '!'+message.content[1:]
+        await bot.delete_message(message)
+        await bot.process_commands(message)
     else:
         if pattern69.search(message.content):
             await bot.add_reaction(message, emoji_letter_map['n'])
